your commit instruction

+++ src/main/java/com/beingjavaguys/model/Obshhee.java
@@ -68,7 +68,7 @@ public class Obshhee implements Serializable {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(long id) {
         this.id = id;
     }
